refactor mobileView to a hook
