+++ lib/hooks.js
@@ -78,7 +78,7 @@ export function useRequireAuth(redirectUrl = '/login') {
 } 
 
 // form handling & validation
-export function useForm (initialValues, callback, validate) {
+export function useForm(initialValues, callback, validate) {
   const [values, setValues] = useState(initialValues)
   const [errors, setErrors] = useState({})
   const [isSubmitting, setIsSubmitting] = useState(false)
@@ -121,4 +121,23 @@ export function useForm (initialValues, callback, validate) {
     values,
     errors,
   }
+}
+
+export function useCheckMobileView() {
+  const [mobileView, setMobileView] = useState(false)
+
+  useEffect(() => {
+    const setResponsiveness = () => {
+      return window.innerWidth < 1024
+        ? setMobileView(true)
+        : setMobileView(false)
+    }
+
+    setResponsiveness()
+    window.addEventListener('resize', () => setResponsiveness())
+
+    return () => window.removeEventListener('resize', () => setResponsiveness())
+  }, [])
+
+  return mobileView
 }
